Move webview to a file
